Get working with ghc-8.2
inline-c has to be limited to 0.5.6.1 because inline-c-0.6.0.0 uses
TH.addForeignFile, which doesn't seem to support separate c++ flags,
which is a problem because Halide requires c++11. Passing -std=c++11 to
to the c-flags (via -optc) does let halide compile but the runtime fails
to compile on MacOS because it complains about using --std=c++11.

There is a proposal that would help solve this problem on Trac:
  https://ghc.haskell.org/trac/ghc/ticket/14298
which allows passing a bytestring containing the object file. However it
looks like this would make it into ghc-8.6 at the earliest.

+++ cbits/halide-hs.cpp
@@ -1,11 +1,6 @@
 // The only Halide header file you need is Halide.h. It includes all of Halide.
 #include "Halide.h"
 
-// We'll also include stdio for printf.
-#include <stdio.h>
-
-/* typedef void CFunc; */
-
 using namespace Halide;
 
 // Turn a pointer of argument points (from haskell) into a vector of
@@ -36,7 +31,7 @@ void delete_param_uint16(Param<uint16_t>* p) { delete p; }
 void delete_param_uint32(Param<uint32_t>* p) { delete p; }
 void delete_param_uint64(Param<uint64_t>* p) { delete p; }
 void delete_param_float(Param<float>* p) { delete p; }
-void delete_param_double(Param<float>* p) { delete p; }
+void delete_param_double(Param<double>* p) { delete p; }
 
 ImageParam* new_image_param(halide_type_t ty, int dim, char* name) {
   ImageParam* img = new ImageParam(ty, dim, name);
